Get rid of one element Arrays
Terraform is not capable of storing a nested Resource in a Resource
because there is no ValueType that maps to an Resource. Instead, they
use a one element Array so that the element type of that array can
be a Resource. This creates a lot of unnecessary arrays and increases
the nesting depth in a bad way.

In this commit, an Array of size one, with an element type that is
a Resource, is considered to be a Resource and the intermediate array
is removed. If the minimum size is zero, the attribute is considered
optional.

The TF marshal/unmarshal uses the Resource and the Schema respectively
to correctly transfer the nested type to/from the needed one element
array.